+++ web/oauth/types.go
@@ -9,11 +9,6 @@ type AccessTokenResponse struct {
 	AccessToken string `json:"access_token"`
 }
 
-type OAuthConfig struct {
-	ClientID string
-	Secret   string
-}
-
 type OauthUserInfo struct {
 	Name       string
 	Email      string
@@ -21,8 +16,8 @@ type OauthUserInfo struct {
 }
 
 type OauthImplementation interface {
-	RequestAccessToken(code, baseurl string, cfg *OAuthConfig) (string, error)
-	RequestUserInfo(access_token string, cfg *OAuthConfig) (*OauthUserInfo, error)
+	RequestAccessToken(code, baseurl string, cfg *types.OAuthConfig) (string, error)
+	RequestUserInfo(access_token string, cfg *types.OAuthConfig) (*OauthUserInfo, error)
 }
 
 type SuccessCallback func(w http.ResponseWriter, user *types.User, new_user bool) error
